Reorganized code within file structure
