+++ multipong/routes.py
@@ -1,0 +1,7 @@
+from flask import render_template
+from multipong import app
+
+
+@app.route('/')
+def index():
+    return render_template("index.html")
